+++ mission-control/app.ts
@@ -13,7 +13,28 @@ var _ = require('lodash');
 const sdpgetter = require("../rtsp-sdp-query")
 const { spawn } = require('child_process');
 
+// Utils
+//-------------
 
+function makeid(length) {
+    var result           = '';
+    var characters       = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    var charactersLength = characters.length;
+    for ( var i = 0; i < length; i++ ) {
+       result += characters.charAt(Math.floor(Math.random() * charactersLength));
+    }
+    return result;
+ }
+
+ function blankMnmsData(d) {
+     let out = JSON.parse(JSON.stringify(d))
+     out.Switches.forEach((s) => {
+        s.Child = null
+        s.Timer = null
+        s.StartTime = null
+     })
+     return out
+ }
 // Options and exports
 //--------------------------------
 
@@ -187,12 +208,17 @@ export = function(LocalOptions) {
                             }
                         }
                     })
-                    Object.keys(Nodes[index].Services).forEach((key) => {
-                        if(!(newValue.Services[key])) {
-                            console.log("Deleting",key)
-                            delete Nodes[index].Services[key]
-                        }
-                    })
+                    if(1) {
+                        Object.keys(Nodes[index].Services).forEach((key) => {
+                            if(!(newValue.Services[key])) {
+                                console.log("Deleting",key)
+                                delete Nodes[index].Services[key]
+                                if(Object.keys(Nodes[index].Services).length == 0) {
+                                    if(Nodes[index].Type && Nodes[index].Type != "switch") Nodes[index].Type = "disconnected"
+                                }
+                            }
+                        })
+                    }
                 }
                 Nodes[index].OtherIPs = newValue.OtherIPs
                 Nodes[index].Macs = newValue.Macs  
@@ -321,6 +347,7 @@ export = function(LocalOptions) {
                                 StartTime: null,
                                 UID: "ddjt" + Date.now() + ((encodeURIComponent(D.IP)))
                             })
+                            db.update({Type: "MnmsData"},blankMnmsData(MnmsData), {upsert: true},(err, newDoc) => { })
                             console.log(MnmsData)
                         }
                     }
@@ -338,44 +365,26 @@ export = function(LocalOptions) {
 
     // db and other services start
     //------------------
-
+    var MnmsData ;
     var Datastore = require('nedb')
     , db = new Datastore({ filename: path.join(__dirname, Options.database), autoload: true });
-
-    var MnmsData = {
-        Type: "MnmsData",
-        Workspace: "Nicolas' test network",
-        CurrentTime: 0,
-        Challenge: "thisisatest",
-        OkSwitches: 0,
-        Switches : [
-     /*           {
-                    Type: "ciscoSG",
-                    IP : "192.168.1.201",
-                    Child: null,
-                    Timer: null,
-                    StartTime: null,
-                    UID: "ddjtzlzégndfe"
-                },
-                {
-                    Type: "ciscoSG",
-                    IP : "192.168.1.129",
-                    Child: null,
-                    Timer: null,
-                    StartTime: null,
-                    UID: "aewtuzhmfdfgh"
-                },*/
-                {
-                    Type: "ciscoSG",
-                    IP : "192.168.1.130",
-                    Child: null,
-                    Timer: null,
-                    StartTime: null,
-                    UID: "bn,héioàtzjrtwrgw"
-                }
-            ]
-    }
-
+    db.find({ Type: "MnmsData"}, (err, docs) => {
+        console.log(docs)
+        if(docs.length==1) {
+            MnmsData = docs[0]
+        }
+        else {
+            MnmsData = {
+                Type: "MnmsData",
+                Workspace: "Mnms - Network Name",
+                CurrentTime: 0,
+                Challenge: makeid(20),
+                OkSwitches: 0,
+                Switches : []
+            }
+        }
+    })
+    
     var ServicesDirectory = {
         cisco_switch: "../cisco-switch/app.js"
     }
@@ -395,7 +404,7 @@ export = function(LocalOptions) {
                 })
             }
             else if(action == "stop") {
-                ServiceOptions.Params.Child.kill()
+                if(ServiceOptions.Params.Child.kill) ServiceOptions.Params.Child.kill()
                 child_info = null;
             }
         }
